refactor(macos): extract album playback initiation from AppModel into AppModel+Playback.swift
Continues the AppModel god-object split. Relocates play(tracks:) — the core
"replace the queue and start playing" entry point — and the album-level actions
(play / shuffle / playNext / addToQueue) into a new `extension AppModel` file.
The favorite-cache stored props (favoriteById, favoriteChangeToken) stay on the
main class under a new "Favorite cache" MARK.

Pure relocation: the 77 moved lines are byte-identical to what was removed.
AppModel.swift drops 4,143 -> 4,063 lines. No promotions needed.

Verified: clean rebuild (rm -rf macos/.build && swift build) + swift test
(889 tests) green.
